Rename validator classes from *Validation to *Validator to be consistent and to use a better noun for objects. Rename validation test classes from *ValidatorTest to *ValidationTest, since they're not really unit tests of the *Validator classes, but rather tests of certain kinds of validation.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=211827852
