+++ java/dagger/internal/codegen/BUILD
@@ -202,17 +202,17 @@ java_library(
     name = "binding_graph_validation",
     srcs = [
         "BindingGraphValidationModule.java",
-        "DependencyCycleValidation.java",
+        "DependencyCycleValidator.java",
         "DependsOnProductionExecutorValidator.java",
-        "DuplicateBindingsValidation.java",
-        "IncompatiblyScopedBindingsValidation.java",
+        "DuplicateBindingsValidator.java",
+        "IncompatiblyScopedBindingsValidator.java",
         "IncorrectlyInstalledBindsMethodsValidator.java",
-        "InjectBindingValidation.java",
-        "MapMultibindingValidation.java",
-        "MissingBindingValidation.java",
-        "NullableBindingValidation.java",
-        "ProvisionDependencyOnProducerBindingValidation.java",
-        "SubcomponentFactoryMethodValidation.java",
+        "InjectBindingValidator.java",
+        "MapMultibindingValidator.java",
+        "MissingBindingValidator.java",
+        "NullableBindingValidator.java",
+        "ProvisionDependencyOnProducerBindingValidator.java",
+        "SubcomponentFactoryMethodValidator.java",
         "Validation.java",
     ],
     plugins = CODEGEN_PLUGINS,
