+++ javatests/dagger/internal/codegen/CanReleaseReferencesValidationTest.java
@@ -28,7 +28,7 @@ import org.junit.runners.JUnit4;
 
 /** Tests for {@link CanReleaseReferencesValidator}. */
 @RunWith(JUnit4.class)
-public final class CanReleaseReferencesValidatorTest {
+public final class CanReleaseReferencesValidationTest {
   @Test
   public void annotatesSourceRetainedAnnotation() {
     JavaFileObject annotation =
